Program now stores history and can execute commands from history.
Checks the command the user tries to run from history to make sure it exists and doesn't try to find locations outwith the size of the array.
User enters !1 to run the first command which is stored in location 0 of the array.

Array is a circular char array with a modular counter, so when it is full the location counter is reset to 0 and commands are replaced from
0 onwards.